move functions into math.ts
