+++ src/components/pages/home.js
@@ -3,7 +3,7 @@ import p5 from 'p5'
 import Matter from 'matter-js'
 import { connect } from 'react-redux'
 
-import { setupFrame, resetPageFrame, setupHome, createMouseConstraint } from '../types'
+import { setupFrame, resetPageFrame, setupHome, createMouseConstraint, renderGroup } from '../types'
 
 const { Engine, World } = Matter
