+++ src/components/types/index.js
@@ -4,4 +4,4 @@ export * from './utils'
 export * from './types'
 export * from './physics'
 export * from './drawing'
-export * from './rendering'
+export * from './utilities/rendering'
