+++ src/components/types/physics.js
@@ -1,14 +1,8 @@
 import Matter from 'matter-js'
-import { ColorBall, Bubble } from './constructors'
-import { getRandomInt, randomColor, desaturateColor } from './'
+// import { ColorBall, Bubble } from './constructors'
+// import { getRandomInt, randomColor, desaturateColor } from './'
 
 const { World, Mouse, MouseConstraint } = Matter
-
-export const addToWorld = (world, instance, group) => {
-  World.add(world, instance.body)
-  group.push(instance)
-  instance.index = group.length - 1
-}
 
 export const createMouseConstraint = (canvas, engine, world, p5) => {
   const mouse = Mouse.create(canvas.elt)
@@ -20,44 +14,44 @@ export const createMouseConstraint = (canvas, engine, world, p5) => {
   World.add(world, mouseConstraint)
 }
 
-export const createColorParticles = (environment) => {
-  const { width } = environment
-  const particleSettings = {
-    x: width * 0.3,
-    y: 10,
-    r: getRandomInt(4, 9),
-    options: {
-      friction: 0,
-      restitution: 0.4,
-      isStatic: false
-    },
-    color: randomColor()
-  }
+// export const createColorParticles = (environment) => {
+//   const { width } = environment
+//   const particleSettings = {
+//     x: width * 0.3,
+//     y: 10,
+//     r: getRandomInt(4, 9),
+//     options: {
+//       friction: 0,
+//       restitution: 0.4,
+//       isStatic: false
+//     },
+//     color: randomColor()
+//   }
 
-  new ColorBall(environment, particleSettings)
-}
+//   new ColorBall(environment, particleSettings)
+// }
 
-export const createBubbles = (environment, button) => {
-  const position = button.body.position
-  const textSize = button.config.textSettings.textSize
+// export const createBubbles = (environment, button) => {
+//   const position = button.body.position
+//   const textSize = button.config.textSettings.textSize
 
-  button.values.forEach(value => {
-    let x = getRandomInt(position.x - 10, position.x + 10)
-    let y = getRandomInt(position.y - 30, position.y - 50)
+//   button.values.forEach(value => {
+//     let x = getRandomInt(position.x - 10, position.x + 10)
+//     let y = getRandomInt(position.y - 30, position.y - 50)
 
-    new Bubble(environment, {
-      x,
-      y,
-      options: {
-        frictionAir: 0.25,
-        restitution: 0.8,
-        isStatic: false
-      },
-      textSettings: {
-        text: value,
-        textSize
-      },
-      color: desaturateColor(environment, randomColor())
-    })
-  })
-}
+//     new Bubble(environment, {
+//       x,
+//       y,
+//       options: {
+//         frictionAir: 0.25,
+//         restitution: 0.8,
+//         isStatic: false
+//       },
+//       textSettings: {
+//         text: value,
+//         textSize
+//       },
+//       color: desaturateColor(environment, randomColor())
+//     })
+//   })
+// }
